refactor: validated and perf tested
