+++ packages/common/src/permissions/types/HubLicense.ts
@@ -1,0 +1,4 @@
+/**
+ * Defines the values for hub license types
+ */
+export type HubLicense = "hub-basic" | "hub-premium" | "enterprise-sites";
