+++ packages/common/src/permissions/types/index.ts
@@ -1,8 +1,8 @@
-export * from "./IEntityPermissionPolicy";
+export * from "./HubLicense";
 export * from "./IAccessResponse";
+export * from "./IEntityPermissionPolicy";
 export * from "./IPermissionPolicy";
 export * from "./IPolicyCheck";
-export * from "./PolicyResponse";
 export * from "./Permission";
-
 export * from "./PlatformPrivilege";
+export * from "./PolicyResponse";
